+++ django_fido/__init__.py
@@ -1,4 +1,4 @@
 """Django application for FIDO protocol."""
 from __future__ import unicode_literals
 
-__version__ = '0.22'
+__version__ = '0.23'